% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actual_cols_used.R
\name{actual_cols_used}
\alias{actual_cols_used}
\title{Title: A function to get the actual columns used in the assay.}
\usage{
actual_cols_used(dat)
}
\arguments{
\item{dat}{A string ("dat_1.dat") if the file is found within the present working directory (pwd) OR a path pointing directly to a ".dat" file.}
}
\value{
Returns a numeric vector denoting the columns used in the assay.
}
\description{
A function that facilitates a users' workflow by helping extract the actual columns used in the assay.
}
\examples{
fpath <- system.file("extdata", "dat_1.dat", package = "normfluodbf", mustWork = TRUE)
acutest <- actual_cols_used(fpath)
}
\author{
Tingwei Adeck
}
