% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_package_namespace}
\alias{add_package_namespace}
\title{Add Package Namespace}
\usage{
add_package_namespace(dir, package)
}
\arguments{
\item{dir}{dir}

\item{package}{package}
}
\value{
modified file
}
\description{
Add Package Namespace
}
\examples{
\dontrun{add_package_namespace(dir, package)}
}
