% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_analysis.R
\name{analyze}
\alias{analyze}
\alias{analyze_ready}
\alias{analyze_ready.default}
\alias{analyze_ready.96well_plate}
\alias{analyze_ready.384well_plate}
\alias{analyze_ready.1536well_plate_t1}
\alias{analyze_ready.1536well_plate_t2}
\title{Analyze}
\usage{
analyze_ready(plate)

\method{analyze_ready}{default}(plate)

\method{analyze_ready}{`96well_plate`}(plate)

\method{analyze_ready}{`384well_plate`}(plate)

\method{analyze_ready}{`1536well_plate_t1`}(plate)

\method{analyze_ready}{`1536well_plate_t2`}(plate)
}
\arguments{
\item{plate}{plate}
}
\value{
plate

plate

plate

plate

plate

plate

plate
}
\description{
Analyze
}
\examples{
\dontrun{analyze_ready(plate)}
}
\concept{analyze}
