% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{average_fluorescence_by_row_cycle}
\alias{average_fluorescence_by_row_cycle}
\title{Capitalize}
\usage{
average_fluorescence_by_row_cycle(plate)
}
\arguments{
\item{plate}{plate}
}
\value{
capital letter
}
\description{
Capitalize
}
\examples{
\dontrun{average_fluorescence_by_row_cycle(plate)}
}
