% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_broken_packages}
\alias{check_broken_packages}
\title{Check Broken Packages}
\usage{
check_broken_packages()
}
\value{
broken packages
}
\description{
Check Broken Packages
}
\concept{dev_helpers}
\concept{utils}
