% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dirs.R
\name{check_dat}
\alias{check_dat}
\title{Title: A function to check for DATs in a directory.}
\usage{
check_dat(pathstring)
}
\arguments{
\item{pathstring}{A string for a path to a directory containing files.}
}
\value{
Returns a list of DAT files or a warning.
}
\description{
A function that facilitates a users' workflow by helping to check for DATs in a directory.
}
\examples{
\dontrun{
fpath <- system.file("extdata", package = "normfluodbf", mustWork = TRUE)
setwd(fpath)
check_dat(getwd())
}
}
\author{
Tingwei Adeck
}
