% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dbfs_pwd.R
\name{check_dbfs_pwd}
\alias{check_dbfs_pwd}
\title{Title: A function to check for DBF files in a directory}
\usage{
check_dbfs_pwd(pathstring)
}
\arguments{
\item{pathstring}{A string that denotes the path to a directory preferably with dbf files}
}
\value{
A list of DBF files if present in the pwd OR a caution to find a folder with DBF files
}
\description{
The function checks for the presence of DBF files in a directory OR demands that the user migrates to one.
}
\examples{
fpath <- system.file("extdata", package = "normfluodbf", mustWork = TRUE)
check_dbfs_pwd(fpath)
}
\author{
Adopted: Tingwei Adeck
}
