% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_utils.R
\name{check_dirt}
\alias{check_dirt}
\title{Check Dirt}
\usage{
check_dirt(plate)
}
\arguments{
\item{plate}{plate}
}
\value{
plate
}
\description{
Check Dirt
}
\seealso{
Other plate_utils: 
\code{\link{get_wells_used}()},
\code{\link{plate_data_summary}()},
\code{\link{platename}},
\code{\link{remove_leading_zero}()},
\code{\link{saveloadutils}},
\code{\link{set_plate_version}()},
\code{\link{type}()}
}
\concept{plate_utils}
