% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_package_usage}
\alias{check_package_usage}
\title{Check package or function Usage}
\usage{
check_package_usage(directory, package_name)
}
\arguments{
\item{directory}{directory}

\item{package_name}{package or string}
}
\value{
use location
}
\description{
Check package or function Usage
}
\examples{
\dontrun{check_package_usage('R','capitalize')}
}
