% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_type.R
\name{childtype}
\alias{childtype}
\alias{child_plate_type}
\alias{child_plate_type.default}
\alias{child_plate_type.normfluodbf_plate}
\title{Child Type}
\usage{
child_plate_type(plate, child_type = NULL)

\method{child_plate_type}{default}(plate, child_type = NULL)

\method{child_plate_type}{normfluodbf_plate}(plate, child_type = NULL)
}
\arguments{
\item{plate}{plate}

\item{child_type}{child type}
}
\value{
plate

plate
}
\description{
Child Type
}
\examples{
\dontrun{child_plate_type()}

}
\concept{childtype}
