% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_evendat.R
\name{clean_evendat}
\alias{clean_evendat}
\title{Title: A function to partially clean dat files with an identified pattern from FLuoStar experiments}
\usage{
clean_evendat(df)
}
\description{
Title: A function to partially clean dat files with an identified pattern from FLuoStar experiments
}
