% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_odddat.R
\name{clean_odddat}
\alias{clean_odddat}
\title{Title: A function to partially clean dat files from FLuostar experiments}
\usage{
clean_odddat(df)
}
\arguments{
\item{df}{A data frame with n number of rows}
}
\value{
A new data frame with most impurities taken out; See the example
}
\description{
The function takes a a dirty data frame and returns a clean data frame with the only exception being that most values in the returned data frame have commas at the end
The data frame that goes into the function comes from reading dat files from FLUOstar experiments
See the example code
}
\note{
This function should work on all types of Fluostar dat files (unlike the clean_evendat() function).
}
\examples{
fpath1 <- system.file("extdata", "dat_1.dat", package = "normfluodbf", mustWork = TRUE)
fpath2 <- system.file("extdata", "dat_2.dat", package = "normfluodbf", mustWork = TRUE)
dat1_df <- read.table(file=fpath1)
dat2_df <- read.table(file=fpath2)
partial_cleaned_dat1 <- clean_odddat(dat1_df)
partial_cleaned_dat1 <- clean_odddat(dat1_df)
}
\author{
Tingwei Adeck
}
