% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comment_out_lines}
\alias{comment_out_lines}
\title{Comment Out}
\usage{
comment_out_lines(input_file, output_file)
}
\arguments{
\item{input_file}{file}

\item{output_file}{file}
}
\value{
file
}
\description{
Comment Out
}
\examples{
\dontrun{comment_out_lines('tests/testthat/test_pipeline.R', 'tests/testthat/test_pipeline.R')}
}
