% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_1}
\alias{dat_1}
\title{dat_1.}
\format{
An object of class \code{data.frame} with 320 rows and 12 columns.
}
\usage{
dat_1
}
\description{
FLUOstar .dat files used for creation of the update and unusable for data immediate data analysis.
}
\keyword{datasets}
