% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_2}
\alias{dat_2}
\title{dat_2.}
\format{
An object of class \code{data.frame} with 320 rows and 12 columns.
}
\usage{
dat_2
}
\description{
FLUOstar .dat files used for creation of the update and unusable for immediate data analysis.
}
\keyword{datasets}
