% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_3}
\alias{dat_3}
\title{dat_3.}
\format{
An object of class \code{data.frame} with 320 rows and 12 columns.
}
\usage{
dat_3
}
\description{
FLUOstar .dat files used for creation of the update and unusable for immediate data analysis.
This file is unique because it validates a major bug fix to ensure that users get the right output.
}
\keyword{datasets}
