% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_4}
\alias{dat_4}
\title{dat_4.}
\format{
An object of class \code{data.frame} with 320 rows and 1 columns.
}
\usage{
dat_4
}
\description{
FLUOstar .dat files used for creation of the update and unusable for immediate data analysis.
This file is unique because it validates a major bug fix to ensure that users get the right output.
}
\keyword{datasets}
