% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_7}
\alias{dat_7}
\title{dat_7.}
\format{
An object of class \code{data.frame} with 105 rows and 3 columns.
}
\usage{
dat_7
}
\description{
FLUOstar .dat files used for creation of the update and unusable for immediate data analysis.
This file is unique because it validates a major bug fix to ensure that users get the right output.
}
\keyword{datasets}
