% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_tidy_dbf.R
\name{decimal_scaling}
\alias{decimal_scaling}
\title{Title: A function that performs decimal scaling (not normalization) of attributes}
\usage{
decimal_scaling(x)
}
\arguments{
\item{x}{Column or attribute values passed into the decimal scaling function}
}
\value{
Attribute(s) with values that have the decimal place moved to facilitate machine learning
}
\description{
Title: A function that performs decimal scaling (not normalization) of attributes
}
\note{
The lapply function is required to apply the function across several columns in a data set.
This is NOT a normalization function because the data still exist on a sliding scale.
This function is here for demonstration purposes and should be used for exercise as it is here for educational purposes for the inventor of the package.
}
\examples{
test_df <- as.data.frame(c(seq(40)))
colnames(test_df) <- "test"
test_df_norm <- lapply(test_df[1:ncol(test_df)], decimal_scaling)
}
\references{
https://www.statology.org/how-to-normalize-data-in-r/
}
\author{
Tingwei Adeck
}
