% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_params.R
\name{defineparams}
\alias{defineparams}
\alias{define_params}
\alias{define_params.default}
\alias{define_params.96well_plate}
\alias{define_params.384well_plate}
\alias{define_params.1536well_plate_t1}
\alias{define_params.1536well_plate_t2}
\alias{set_default_params}
\title{Define Plate Parameters}
\usage{
define_params(plate)

\method{define_params}{default}(plate)

\method{define_params}{`96well_plate`}(plate)

\method{define_params}{`384well_plate`}(plate)

\method{define_params}{`1536well_plate_t1`}(plate)

\method{define_params}{`1536well_plate_t2`}(plate)

set_default_params(plate)
}
\arguments{
\item{plate}{plate}
}
\value{
default params

default params

default params

default params

default params

plate
}
\description{
Define Plate Parameters
}
\examples{
\dontrun{define_params(plate)}
}
\concept{defineparams}
