% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_status.R
\name{definestatus}
\alias{definestatus}
\alias{define_status}
\alias{define_status.default}
\alias{define_status.96well_plate}
\alias{define_status.384well_plate}
\alias{define_status.1536well_plate_t1}
\alias{define_status.1536well_plate_t2}
\alias{set_default_status}
\alias{update_status_list}
\alias{get_status_value}
\title{Define Plate Status}
\usage{
define_status(plate)

\method{define_status}{default}(plate)

\method{define_status}{`96well_plate`}(plate)

\method{define_status}{`384well_plate`}(plate)

\method{define_status}{`1536well_plate_t1`}(plate)

\method{define_status}{`1536well_plate_t2`}(plate)

set_default_status(plate)

update_status_list(plate)

get_status_value(plate, index)
}
\arguments{
\item{plate}{plate}

\item{index}{index}
}
\value{
plate

status

status

status

status

status

status

plate

plate

plate
}
\description{
Define Plate Status
}
\examples{
\dontrun{define_steps(plate)}
}
\concept{definestatus}
