% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{demo_background}
\alias{demo_background}
\title{Run Demo Background}
\usage{
demo_background(
  host = getOption("shiny.host", "127.0.0.1"),
  appDir = system.file("shiny/demo", package = "normfluodbf")
)
}
\arguments{
\item{host}{localhost}

\item{appDir}{dir}
}
\description{
Run Demo Background
}
\examples{
\dontrun {
demo_background(...)}
}
