% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{detectoutliers}
\alias{detectoutliers}
\alias{detect_outliers_time_cn}
\alias{detect_outliers_cn}
\title{Detect Outliers}
\usage{
detect_outliers_time_cn(plate, data)

detect_outliers_cn(plate, data)
}
\arguments{
\item{plate}{plate}

\item{data}{data}
}
\value{
data frame

data frame

data frame
}
\description{
Detect Outliers
}
\examples{
\dontrun{detect_outliers_time_cn(plate, data)}
}
\concept{detectoutliers}
