% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_value_swap.R
\name{df_value_swap}
\alias{df_value_swap}
\title{Title: A function for data frame value replacement}
\usage{
df_value_swap(df, row_num, col_num, replacement_value)
}
\arguments{
\item{df}{A clean data frame}

\item{row_num}{row number for replacement value insertion}

\item{col_num}{column number for replacement value insertion}

\item{replacement_value}{value used to replace original df value at chosen location}
}
\value{
Return a data frame with a swapped value.
}
\description{
Swaps values at a specific location with an input replacement value
}
\examples{
fpath <- system.file("extdata", "dat_1.dat", package = "normfluodbf", mustWork = TRUE)
dat_df <- read.table(file=fpath)
comma_dat <- clean_odddat(dat_df)
nocomma_dat <- comma_cleaner(comma_dat)
resampled_scaled <- resample_dat_scale(nocomma_dat, tnp=3, cycles=40)
nofunions <- df_value_swap(resampled_scaled, 2, 2, 2^16)
fluor_threshold_check(resampled_scaled)
}
