% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{find_random_port}
\alias{find_random_port}
\title{Random Port Normfluodbf Style}
\usage{
find_random_port()
}
\value{
random port
}
\description{
Random Port Normfluodbf Style
}
\details{
Recursively find a random port that does not fall in the unsafe group
Added some more unsafe ports for like PostgresDb (5432), MySQL (3306), StreamlitApp (8501),
ngrok (4040), pinggy (4300), Flask (5000), Django (8000) and ReactApp (3000).
Learnt a thing or two from Dean with recursion (Really Cool!!!).
}
