% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_data.R
\name{formatplatedata}
\alias{formatplatedata}
\alias{format_plate_data}
\alias{format_plate_data.default}
\alias{format_plate_data.96well_plate}
\alias{format_plate_data.384well_plate}
\alias{format_plate_data.1536well_plate_t1}
\alias{format_plate_data.1536well_plate_t2}
\title{Format Plate Data}
\usage{
format_plate_data(plate)

\method{format_plate_data}{default}(plate)

\method{format_plate_data}{`96well_plate`}(plate)

\method{format_plate_data}{`384well_plate`}(plate)

\method{format_plate_data}{`1536well_plate_t1`}(plate)

\method{format_plate_data}{`1536well_plate_t2`}(plate)
}
\arguments{
\item{plate}{plate}
}
\value{
plate

plate

plate

plate

plate

plate

plate
}
\description{
Format Plate Data
}
\examples{
\dontrun{format_plate_data(plate)}
}
\concept{formatplatedata}
