% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_utils.R
\name{get_wells_used}
\alias{get_wells_used}
\title{Wells Used}
\usage{
get_wells_used(pl_data)
}
\arguments{
\item{pl_data}{data}
}
\value{
wells used
}
\description{
Wells Used
}
\examples{
\dontrun{get_wells_used(data)}
}
\seealso{
Other plate_utils: 
\code{\link{check_dirt}()},
\code{\link{plate_data_summary}()},
\code{\link{platename}},
\code{\link{remove_leading_zero}()},
\code{\link{saveloadutils}},
\code{\link{set_plate_version}()},
\code{\link{type}()}
}
\concept{plate_utils}
