% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plot_triplets.R
\name{gg_plot_triplets}
\alias{gg_plot_triplets}
\title{Title: The visualizer function}
\usage{
gg_plot_triplets(df, x, y_list, xlim, ylim)
}
\arguments{
\item{df}{clean normalized dat data frame}

\item{x}{Cycle_No or Time}

\item{y_list}{sample types (tnp = test, negative, positive)}

\item{xlim}{x-axis range (0: number of cycles)}

\item{ylim}{y-axis range (normalization range); 0:1 in this case}
}
\value{
A normalized data frame with the x-variable (Cycle_No), ready for analysis
}
\description{
Visualizes liposome flux assay experiments for all three sample types
}
\note{
This function is not very modular so use with caution. It should work just fine in most cases but here for demonstration.
}
\examples{
fpath <- system.file("extdata", "dat_1.dat", package = "normfluodbf", mustWork = TRUE)
fpath_dbf <- system.file("extdata", "liposomes_214.dbf", package = "normfluodbf", mustWork = TRUE)
n <- c('A','B','C')
normalized_fluo_dat <- normfluodat(dat=fpath, tnp = 3, cycles = 40, n)
normalized_dbf_scale100 <- norm_tidy_dbf(file=fpath_dbf, norm_scale = 'z-score')
yvars <- c("A1","B1","C1")
yvars_dbf <- c('A01','A02','A03')
xvar <- c("Cycle_Number")
color <- c("Test","Negative Control","Positive Control")
xl <- c(0,40)
yl <- c(0,1)
yl_dbf <- c(0,100)
gg_plot_triplets(normalized_fluo_dat,x=xvar,y_list=yvars,xlim=xl,ylim=yl)
gg_plot_triplets(normalized_dbf_scale100,x=xvar,y_list=yvars_dbf,xlim=xl,ylim=yl_dbf)
}
\author{
Tingwei Adeck
}
