% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%there\%}
\alias{\%there\%}
\title{The \%there\% operator}
\usage{
dfile \%there\% dirpath
}
\arguments{
\item{dfile}{file}

\item{dirpath}{directory}
}
\value{
logical
}
\description{
The \%there\% operator
}
\examples{
\dontrun{
fpath <- system.file("extdata", package = "normfluodbf", mustWork = TRUE)
"dat_1.dat" \%there\% fpath}
}
\concept{operators}
