% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normfluodbf-package.R
\name{internals}
\alias{internals}
\alias{normfluodbf_data}
\alias{inspect_workspace}
\alias{get_package_data}
\alias{check_cols_class}
\alias{search_pkg}
\alias{time_it}
\title{Internal Troubleshoot}
\usage{
normfluodbf_data()

inspect_workspace()

get_package_data(name)

check_cols_class(df)

search_pkg(pkg)

time_it(expr)
}
\arguments{
\item{name}{name}

\item{df}{data frame}

\item{pkg}{search value}

\item{expr}{expr}
}
\value{
internals

all data

workspace

pkg data

classes

package

duration
}
\description{
Internal Troubleshoot
}
\examples{
\dontrun{
get_package_data('iris')}
}
\keyword{internal}
