% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{isnormalized}
\alias{isnormalized}
\alias{is_normalized}
\title{Is Normalized}
\usage{
is_normalized(data, type = c("min-max", "z-score", "hundred"))
}
\arguments{
\item{data}{type}

\item{type}{type}
}
\value{
boolean

boolean
}
\description{
Is Normalized
}
\examples{
\dontrun{is_normalized(data,type)}
}
