% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{launch}
\alias{launch}
\title{Run the shiny App
checking dependencies in R code ... WARNING
In addition to the functions provided in this package, the \code{normfluodbf} package
also provides an interactive tool that can be used to analyze liposome flux assay data
more easily. The tool will be launched in a web browser.}
\usage{
launch()
}
\description{
Run the shiny App
checking dependencies in R code ... WARNING
In addition to the functions provided in this package, the \code{normfluodbf} package
also provides an interactive tool that can be used to analyze liposome flux assay data
more easily. The tool will be launched in a web browser.
}
