% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liposomes_214}
\alias{liposomes_214}
\title{liposomes_214.}
\format{
An object of class \code{data.frame} with 11 rows and 52 columns.
}
\usage{
liposomes_214
}
\description{
FLUOstar .dbf file in wide format and unable to use to data analysis.
}
\keyword{datasets}
