% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liposomes_216}
\alias{liposomes_216}
\title{liposomes_216.}
\format{
An object of class \code{data.frame} with 8 rows and 52 columns.
}
\usage{
liposomes_216
}
\description{
FLUOstar .dbf file in wide format and unable to use for data analysis.
}
\keyword{datasets}
