% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liposomes_218}
\alias{liposomes_218}
\title{liposomes_218.}
\format{
An object of class \code{data.frame} with 11 rows and 52 columns.
}
\usage{
liposomes_218
}
\description{
FLUOstar .dbf file in wide format and unable to use for data analysis.
}
\keyword{datasets}
