% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liposomes_221}
\alias{liposomes_221}
\title{liposomes_221.}
\format{
An object of class \code{data.frame} with 38 rows and 52 columns.
}
\usage{
liposomes_221
}
\description{
FLUOstar .dbf file in wide format and unable to use for data analysis.
}
\keyword{datasets}
