% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liposomes_227}
\alias{liposomes_227}
\title{liposomes_227.}
\format{
An object of class \code{data.frame} with 29 rows and 52 columns.
}
\usage{
liposomes_227
}
\description{
FLUOstar .dbf file in wide format and unable to use for data analysis.
}
\keyword{datasets}
