% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_data.R
\name{loadplatedata}
\alias{loadplatedata}
\alias{load_plate_data}
\alias{load_plate_data<-}
\title{Load Plate Data}
\usage{
load_plate_data(plate)

load_plate_data(plate) <- value
}
\arguments{
\item{plate}{plate}

\item{value}{data}
}
\value{
plate

plate

plate
}
\description{
Load Plate Data
}
\examples{
\dontrun{load_plate_data(plate,value = data)}
}
\concept{loadplatedata}
