% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_meta.R
\name{loadplatemeta}
\alias{loadplatemeta}
\alias{load_plate_meta}
\alias{load_plate_meta<-}
\title{Load Plate Meta}
\usage{
load_plate_meta(plate)

load_plate_meta(plate) <- value
}
\arguments{
\item{plate}{plate}

\item{value}{metadata}
}
\value{
plate

plate

plate
}
\description{
Load Plate Meta
}
\examples{
\dontrun{load_plate_meta(plate, meta)}

}
\concept{loadplatemeta}
