% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_tidy_dbf.R
\name{log_transformation}
\alias{log_transformation}
\title{Title: A function that performs log transformation of data}
\usage{
log_transformation(x)
}
\arguments{
\item{x}{Column or attribute values passed into the Z-standardization function}
}
\value{
Log transformed attributes
}
\description{
Title: A function that performs log transformation of data
}
\note{
The lapply function is required to apply the function across several columns in a data set.
}
\examples{
test_df <- as.data.frame(c(seq(40)))
colnames(test_df) <- "test"
test_df_norm <- lapply(test_df[1:ncol(test_df)], log_transformation)
}
\references{
https://www.statology.org/how-to-normalize-data-in-r/
}
\author{
Tingwei Adeck
}
