% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntdbf_subordinates.R
\name{min_max_norm_percent}
\alias{min_max_norm_percent}
\title{Title: Min-Max normalization on a 0-100 scale.}
\usage{
min_max_norm_percent(x)
}
\arguments{
\item{x}{Attribute value(s).}
}
\value{
A normalized value (between 0 and 100) when applied to a single value or a normalized attribute with values between 0 and 100.
}
\description{
Title: Min-Max normalization on a 0-100 scale.
}
\note{
The lapply function is required to apply this function across several attributes.
}
\examples{
test_df <- as.data.frame(c(seq(40)))
colnames(test_df) <- "test"
test_df_norm <- lapply(test_df[1:ncol(test_df)], min_max_norm_percent)

}
\references{
https://www.statology.org/how-to-normalize-data-in-r/
}
\author{
Tingwei Adeck
}
