% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_data.R
\name{modifyplatedata}
\alias{modifyplatedata}
\alias{modify_plate_data}
\alias{modify_plate_data.default}
\alias{modify_plate_data.96well_plate}
\alias{modify_plate_data.384well_plate}
\alias{modify_plate_data.1536well_plate_t1}
\alias{modify_plate_data.1536well_plate_t2}
\title{Modify Plate Data}
\usage{
modify_plate_data(plate)

\method{modify_plate_data}{default}(plate)

\method{modify_plate_data}{`96well_plate`}(plate)

\method{modify_plate_data}{`384well_plate`}(plate)

\method{modify_plate_data}{`1536well_plate_t1`}(plate)

\method{modify_plate_data}{`1536well_plate_t2`}(plate)
}
\arguments{
\item{plate}{plate}
}
\value{
plate

plate

plate

plate

plate

plate

plate
}
\description{
Modify Plate Data
}
\examples{
\dontrun{modify_plate_meta(plate)}
}
\concept{modifyplatedata}
