% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{move_file}
\alias{move_file}
\title{Move File}
\usage{
move_file(source_path, destination_path)
}
\arguments{
\item{source_path}{src}

\item{destination_path}{dest}
}
\value{
kinetic file
}
\description{
Move File
}
\examples{
\dontrun{
source_file <- "path/to/source/file.txt"
destination_file <- "path/to/destination/file.txt"
move_file(source_file, destination_file)
move_file("/home/alphaprime7/Documents/Wip/R/PkgDev/pdf","/home/alphaprime7/Documents/Wip/R/PkgDev/R/pdf")}
}
