% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{extra}

\item{plotlist}{list}

\item{file}{file}

\item{cols}{cols}

\item{layout}{layout}
}
\value{
grid plot
}
\description{
Multiplot
}
