% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfd_tracker.R
\name{nfd_tracker}
\alias{nfd_tracker}
\title{Title: Normfluodbf tracker}
\usage{
nfd_tracker(package, period = NULL, plot = NULL)
}
\arguments{
\item{package}{package name as string (example: package = 'normfluodbf')}

\item{period}{takes 'last-month' or 'last-week'; default is NULL}

\item{plot}{takes 'cum' or 'daily'; default is null}
}
\value{
A data frame of cumulative downloads or a data frame of daily downloads
}
\description{
A simple tracker for the package. A fun addition to the package.
}
\note{
The tracker is not modular so only works for normfluodbf hence in the package
}
\examples{
nfd_tracker(package = 'normfluodbf', period = 'last-month', plot = 'cum')
}
\author{
Tingwei Adeck
}
