% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normfluodat.R
\name{norm_applier}
\alias{norm_applier}
\title{Title: A normalization applier built on lapply.}
\usage{
norm_applier(df, norm_scale = c("one", "hundred", "z-score", "raw", "decimal"))
}
\arguments{
\item{df}{A data frame.}

\item{norm_scale}{This parameter takes sub-parameters: 'raw’ , hundred’ , 'one’ , 'z-score' , or 'decimal’ ,
which denotes the normalization type or scale.}
}
\value{
A data frame with attribute values obtained from the applied function using lapply.
}
\description{
Applies a function over a list of attributes.
}
\examples{
test_df <- as.data.frame(c(seq(40)))
colnames(test_df) <- "test"
test_df_norm <- norm_applier(test_df,norm_scale = 'one')
}
