% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizebywell}
\alias{normalizebywell}
\alias{normalize_by_well}
\alias{normalize_by_well.default}
\alias{normalize_by_well.96well_plate}
\alias{normalize_by_well.384well_plate}
\alias{normalize_by_well.1536well_plate_t1}
\alias{normalize_by_well.1536well_plate_t2}
\title{Normalize by Well}
\usage{
normalize_by_well(plate)

\method{normalize_by_well}{default}(plate)

\method{normalize_by_well}{`96well_plate`}(plate)

\method{normalize_by_well}{`384well_plate`}(plate)

\method{normalize_by_well}{`1536well_plate_t1`}(plate)

\method{normalize_by_well}{`1536well_plate_t2`}(plate)
}
\arguments{
\item{plate}{plate}
}
\value{
plate

plate

plate

plate

plate

plate

plate
}
\description{
Normalize by Well
}
\examples{
\dontrun{normalize_by_well(plate)}
}
\concept{normalizebywell}
