% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normfluodbf_utils.R
\name{normalizingagents}
\alias{normalizingagents}
\alias{min_max_norm}
\alias{min_max_norm_df}
\alias{min_max_norm_percent}
\alias{min_max_norm_percent_df}
\alias{norm_z}
\alias{norm_z_df}
\alias{decimal_scaling}
\alias{decimal_scaling_df}
\alias{log_transformation}
\alias{roundfluor}
\title{Normalizing Agents}
\usage{
min_max_norm(x)

min_max_norm_df(df)

min_max_norm_percent(x)

min_max_norm_percent_df(df)

norm_z(x)

norm_z_df(df)

decimal_scaling(x)

decimal_scaling_df(df)

log_transformation(x)

roundfluor(x)
}
\arguments{
\item{x}{value(s)}
}
\value{
A normalized value when applied to a single value or a normalized attribute with values between the normalizing range.

normalized value (0-1)

normalized value (0-1)

normalized value (0-100)

normalized value (0-100)

normalized value (Z = N (0,1))

normalized value (Z = N (0,1))

normalized value

normalized value

log value

rounded value
}
\description{
Normalizing Agents
}
\examples{
\dontrun{
test_df <- as.data.frame(c(seq(40)))
colnames(test_df) <- "test"
test_df_norm <- lapply(test_df[1:ncol(test_df)], min_max_norm)
}
}
\seealso{
Other normfluodbf_utils: 
\code{\link{fluorthresholdcheck}}
}
\concept{normfluodbf_utils}
