% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normfluodat.R
\name{normfluodat}
\alias{normfluodat}
\title{Title: The root function that returns a normalized data frame with the Cycle No ready for analysis}
\usage{
normfluodat(dat, tnp, cycles, rows_used = NULL, cols_used = NULL)
}
\arguments{
\item{dat}{directory to the users FLUOstar dat file}

\item{tnp}{Stands for test,negative,positive (sample types); the number should match the number of sample types in the plate reader even if repeating a sample type}

\item{cycles}{The number of cycles chosen by the researcher. In the case of this package 40 is the standard but ensure to have the right number of samples}

\item{rows_used}{A character vector of the rows used, eg n = c('A','B','C')}

\item{cols_used}{A numeric vector of the columns used, eg m = c(2,4,6)}
}
\value{
A normalized data frame with the x-variable (Cycle_No), ready for analysis
}
\description{
Input a dat file (dat file directory) and required parameters and BOOM the researcher has a normalized data frame ripe and ready for clean analysis
}
\note{
This is the MAIN function and stands alone but is dependent on the subordinate functions. If the user understands what they are doing this is all they need.
}
\examples{
fpath <- system.file("extdata", "dat_1.dat", package = "normfluodbf", mustWork = TRUE)
n <- c('A','B','C')
normalized_fluo_dat <- normfluodat(dat=fpath, tnp = 3, cycles = 40, n)
}
\author{
Tingwei Adeck
}
