% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normfluodbfcomms}
\alias{normfluodbfcomms}
\alias{normfluodbf_warn_msg}
\alias{normfluodbf_stop_msg}
\alias{normfluodbf_msg_msg}
\alias{msg}
\title{Directory Utils}
\usage{
normfluodbf_warn_msg(x)

normfluodbf_stop_msg(x)

normfluodbf_msg_msg(x)

msg(...)
}
\arguments{
\item{x}{dev message}

\item{...}{dots}
}
\value{
custom messages
}
\description{
Directory Utils
}
\note{
Overwrite with \code{options(normfluodbf.verbose = FALSE)}.
Write a message to the user if the \code{normfluodbf.verbose} option is on.
}
\examples{
\dontrun{
fpath <- system.file("extdata", package = "normfluodbf", mustWork = TRUE)
normfluodbf_warn_msg(x = 'life is cool')
normfluodbf_msg_msg(x = 'Do some java script and make GUI apps because you earned it')
normfluodbf_stop_msg(x = 'Dont Move Forward')}
}
\seealso{
Other dirutils: 
\code{\link{dirutils}},
\code{\link{sampledata}}
}
\concept{dirutils}
\keyword{internal}
