% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normfluordbf.R
\name{normfluordbf}
\alias{normfluordbf}
\title{Title: Tidy and Normalize .dbf files obtained from experiments using the FLUOstar microplate reader.}
\usage{
normfluordbf(file = NULL, fun = NA, ...)
}
\arguments{
\item{file}{A string ("x.dbf) or path directly pointing to a .dbf file}

\item{fun}{A variable defined as NA, used for boolean expressions}

\item{...}{A sequence of dots}
}
\value{
Normalized dataframe with a Time and Cycle_No column
}
\description{
Normalize a tidy dbf data frame. Specific to FLUOstar dirty dbf files. Version 2 will also address .dat files.
}
\note{
Simply renaming norm_tidy_dbf to a name that I like more. Users can continue with the old name but this is a better name in my opinion.
}
\examples{
fpath <- system.file("extdata", "liposomes_214.dbf", package = "normfluodbf", mustWork = TRUE)
normalized_dbf <- normfluordbf(file=fpath)
}
