% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normfluordbf_legacy.R
\name{normfluordbf_legacy}
\alias{normfluordbf_legacy}
\title{Title: Cleans and Normalizes ".dbf" files obtained from experiments using the FLUOstar microplate reader.}
\usage{
normfluordbf_legacy(file = NULL, fun = NA, ...)
}
\arguments{
\item{file}{A string ("liposomes_xxx.dbf") if the file is found within the present working directory (pwd) OR a path pointing directly to a ".dbf" file, from FLUOstar experiments.}

\item{fun}{A variable defined as NA, used for boolean expressions or manipulation.}

\item{...}{A container object that can be used to capture extra variables if needed.}
}
\value{
A normalized data frame with an appended "Cycle_Number" attribute.
}
\description{
Input the path to a ".dbf" file obtained from the FLUOstar microplate (usually a 96-well microplate) reader; this function will create a data frame, clean the data frame, normalize the data frame, append a "Cycle_Number" column and return a data frame that is ready for analysis.
Most importantly, this function is a single_step function.
Also, the function can be extended to other ".dbf" files if they follow the format for which this function was designed; this is totally at the users' discretion.
}
\note{
Re-nomenclature of norm_tidy_dbf to a more appropriate name that facilitates function utilization. Users can continue with the old name ("norm_tidy_dbf") but this is a better name in my opinion.
This function is the first of the main functions in this package and is a dumbed down version of norm_tidy_dbf and normfluordbf. It is here as a reminder of where the inventor started.
This function needs to be included to ensure the continuum of education and it is the 'ROOT'.
normfluordbf_legacy will be equivalent to norm_tidy_dbf_legacy (this does not exist).
}
\examples{
fpath <- system.file("extdata", "liposomes_214.dbf", package = "normfluodbf", mustWork = TRUE)
normalized_dbf <- normfluordbf_legacy(file=fpath)
}
