% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgTracker.R
\name{package_trackeR}
\alias{package_trackeR}
\title{Title: An adopted function to track normfluodbf}
\usage{
package_trackeR(packages)
}
\arguments{
\item{packages}{A string == name of the package}
}
\value{
A tibble showing download statistics per month, alongside a plot for visualization.
}
\description{
A function that tracks normfluodbf statistics
}
\examples{
package_trackeR(c("normfluodbf"))
}
\seealso{
\code{\link[=nfd_tracker]{nfd_tracker()}}
}
\author{
Gary Hutson (in R Bloggers)

Adopted: Tingwei Adeck
}
