% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_type.R
\name{parenttype}
\alias{parenttype}
\alias{parent_plate_type}
\alias{parent_plate_type.normfluodbf_plate}
\alias{parent_plate_type.default}
\alias{parent_plate_type.96well_plate}
\alias{parent_plate_type.384well_plate}
\alias{parent_plate_type.1536well_plate_t1}
\alias{parent_plate_type.1536well_plate_t2}
\title{Parent Type}
\usage{
parent_plate_type(plate, type = NULL)

\method{parent_plate_type}{normfluodbf_plate}(plate, type = NULL)

\method{parent_plate_type}{default}(plate, type = NULL)

\method{parent_plate_type}{`96well_plate`}(plate, type = NULL)

\method{parent_plate_type}{`384well_plate`}(plate, type = NULL)

\method{parent_plate_type}{`1536well_plate_t1`}(plate, type = NULL)

\method{parent_plate_type}{`1536well_plate_t2`}(plate, type = NULL)
}
\arguments{
\item{plate}{plate}

\item{type}{parent type}
}
\value{
plate

plate

plate

plate

plate
}
\description{
Parent Type
}
\examples{
\dontrun{parent_plate_type()}
}
\concept{parenttype}
