% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{plate}
\alias{plate}
\alias{empty_plate}
\alias{new_plate}
\alias{init_plate}
\alias{setup_plate}
\alias{reset_plate}
\alias{use_setup_plate}
\alias{use_initialized_plate}
\title{Plate}
\usage{
empty_plate()

new_plate()

init_plate(plate = NULL, type = NULL, child_type = NULL)

setup_plate(plate, ...)

reset_plate(plate)

use_setup_plate()

use_initialized_plate()
}
\arguments{
\item{plate}{plate}

\item{type}{parent type}

\item{child_type}{child type}
}
\value{
plate

plate

plate

plate

plate

plate

plate

plate
}
\description{
Plate
}
\examples{
\dontrun{
empty_plate()
new_plate()
init_plate()
setup_plate(plate)
reset_plate()
}
}
\concept{plate}
