% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_utils.R
\name{plate_data_summary}
\alias{plate_data_summary}
\title{Plate Data Summary}
\usage{
plate_data_summary(plate)
}
\arguments{
\item{plate}{plate}
}
\value{
sprintf string
}
\description{
Plate Data Summary
}
\seealso{
Other plate_utils: 
\code{\link{check_dirt}()},
\code{\link{get_wells_used}()},
\code{\link{platename}},
\code{\link{remove_leading_zero}()},
\code{\link{saveloadutils}},
\code{\link{set_plate_version}()},
\code{\link{type}()}
}
\concept{plate_utils}
