% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_type.R, R/normfluodbf-package.R
\docType{data}
\name{plate_types_tbl}
\alias{plate_types_tbl}
\alias{plate_types}
\alias{plate_types_vector}
\title{Plate Types Tibble}
\format{
An object of class \code{list} of length 5.
}
\usage{
plate_types_tbl()

plate_types

plate_types_vector()
}
\value{
A tibble

A list

A Vector
}
\description{
Plate Types Tibble

Plate Types List

Plate Types Vector

Plate Types Global
}
\details{
The list equivalent of the tibble from \code{plate_types_tbl}.

The vector equivalent of the tibble from \code{plate_types_tbl}.
}
\seealso{
\code{\link[=plate_types]{plate_types()}}
}
\keyword{datasets}
