% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_type.R
\docType{data}
\name{plate_types_list}
\alias{plate_types_list}
\alias{plate_types}
\title{Plate Types List}
\format{
An object of class \code{list} of length 5.
}
\usage{
plate_types
}
\value{
A list
}
\description{
Plate Types List
}
\details{
The list equivalent of the tibble from \code{plate_types_tbl}.
}
\keyword{datasets}
