% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_type.R
\name{plate_types_tbl}
\alias{plate_types_tbl}
\title{Plate Types Tibble}
\usage{
plate_types_tbl()
}
\value{
A tibble
}
\description{
Plate Types Tibble
}
\seealso{
\code{\link[=plate_types]{plate_types()}}
}
