% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_type.R
\name{plate_types_vector}
\alias{plate_types_vector}
\title{Plate Types Vector}
\usage{
plate_types_vector()
}
\value{
A Vector
}
\description{
Plate Types Vector
}
\details{
The vector equivalent of the tibble from \code{plate_types_tbl}.
}
