% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_data.R
\name{platedata}
\alias{platedata}
\alias{plate_data}
\title{Plate Data}
\usage{
plate_data(file, tnp = NULL, cycles = NULL, rows_used = NULL, ...)
}
\arguments{
\item{file}{file}

\item{tnp}{tnp}

\item{cycles}{cycles}

\item{rows_used}{rows_used}

\item{...}{dots}
}
\value{
plate data

plate data
}
\description{
Plate Data
}
\examples{
\dontrun{plate_data(file, tnp, cycles, rows_used = c(A,B,C), norm_scale = 'raw')}
}
\concept{platedata}
