% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_meta.R
\name{platemeta}
\alias{platemeta}
\alias{plate_meta}
\alias{plate_meta.default}
\alias{plate_meta.96well_plate}
\alias{plate_meta.384well_plate}
\alias{plate_meta.1536well_plate_t1}
\alias{plate_meta.1536well_plate_t2}
\title{Plate Meta}
\usage{
plate_meta(plate, num_wells)

\method{plate_meta}{default}(plate, num_wells = 96L)

\method{plate_meta}{`96well_plate`}(plate, num_wells = 96L)

\method{plate_meta}{`384well_plate`}(plate, num_wells = 384L)

\method{plate_meta}{`1536well_plate_t1`}(plate, num_wells = 1536L)

\method{plate_meta}{`1536well_plate_t2`}(plate, num_wells)
}
\arguments{
\item{plate}{plate}

\item{num_wells}{number of wells}
}
\value{
plate

plate

plate

plate

plate

plate

plate
}
\description{
Plate Meta
}
\examples{
\dontrun{plate_meta(plate, num_wells = 96L)}

}
\concept{platemeta}
