% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_utils.R
\name{platename}
\alias{platename}
\alias{name}
\alias{name<-}
\title{Plate Name}
\usage{
name(plate)

name(plate) <- value
}
\arguments{
\item{plate}{plate}

\item{value}{value}
}
\value{
plate

plate
}
\description{
Plate Name
}
\seealso{
Other plate_utils: 
\code{\link{check_dirt}()},
\code{\link{get_wells_used}()},
\code{\link{plate_data_summary}()},
\code{\link{remove_leading_zero}()},
\code{\link{saveloadutils}},
\code{\link{set_plate_version}()},
\code{\link{type}()}
}
\concept{plate_utils}
