% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plotutils}
\alias{plotutils}
\alias{reverse_row}
\alias{row_levels_layout}
\alias{row_levels_grid}
\alias{column_levels}
\alias{column_levels_layout}
\alias{column_levels_grid}
\alias{row_levels}
\alias{get_x_var}
\alias{get_y_var}
\alias{convert_x_one_label}
\alias{convert_x_two_label}
\alias{get_x_label}
\alias{add_custom_legend}
\alias{remove_legend}
\alias{reset_grid}
\alias{save_plot}
\alias{plot_canvas}
\alias{plot_in_well}
\alias{plot_grid}
\alias{plot_grid_with_forcats}
\alias{plot_grid_alt.one}
\alias{plot_grid_alt.two}
\alias{plot_grid_alt.three}
\alias{plot_superimpose}
\alias{plot_superimpose_advanced}
\alias{plot_plate_layout}
\alias{plot_return_plot.obj}
\alias{plot_side_by_side}
\alias{plotly_plot}
\alias{plot_dev_deprecated}
\alias{plot_dev}
\alias{plot_dev_with_custom_legends}
\alias{plate_wells_grid}
\title{Adeckstats Plots}
\usage{
reverse_row(vect)

row_levels_layout(plate)

row_levels_grid(rows)

column_levels(plate)

column_levels_layout(plate)

column_levels_grid(cols)

row_levels(plate)

get_x_var(data)

get_y_var(data)

convert_x_one_label(xlabel)

convert_x_two_label(xlabel)

get_x_label(xvar)

add_custom_legend(df, legend_values)

remove_legend(plate)

reset_grid(plate, grid_plot)

save_plot(name)

plot_canvas(plate)

plot_in_well(plt.obj, data, no_annotations = FALSE)

plot_grid(data, plt.obj)

plot_grid_with_forcats(base_plot, data)

plot_grid_alt.one(base_plot, data)

plot_grid_alt.two(base_plot, data)

plot_grid_alt.three(base_plot, data)

plot_superimpose(
  data,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  title = NULL,
  legend_labels = NULL
)

plot_superimpose_advanced(
  data,
  plt.obj,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  legend_labels = NULL
)

plot_plate_layout(plate)

plot_return_plot.obj(data, plt.obj)

plot_side_by_side(
  data,
  plt.obj,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  title = NULL,
  legend_labels = NULL
)

plotly_plot(plate, title = NULL, ylab = NULL, xlab = NULL)

plot_dev_deprecated(
  df,
  wells,
  x = NULL,
  well_colors = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  title = NULL
)

plot_dev(
  df,
  wells,
  x = NULL,
  well_colors = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  title = NULL
)

plot_dev_with_custom_legends(
  df,
  wells,
  legend_labels = NULL,
  x = NULL,
  well_colors = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  title = NULL
)

plate_wells_grid(num_wells = NULL)
}
\arguments{
\item{...}{dots}
}
\value{
plot utils

plot utils

plot utils

plot utils

plot utils

plot utils

plot utils

plot utils

plot utils

plot utils

plot utils

plot utils

plot utils

plot utils

plot utils

plot utils

plot utils

ggplot list

ggplot list

ggplot list

ggplot list

ggplot list

ggplot list

ggplot list

ggplot list

ggplot list

ggplot list

ggplot list

ggplot list

ggplot list

ggplot list

ggplot list

ggplot list

ggplot list
}
\description{
Adeckstats Plots
}
\note{
Just one of those functions I have found no use for

purely subordinate because of the plt.obj found in the plot function

The most useless function here but learnt plotly a little.

Just simple grids with no functionality. No time to delve deeper but a good primer.
}
\examples{
\dontrun {
reverse_row(vect = vector)
column_levels(plate = plate)
row_levels(plate = plate)
base_plot(plate = plate)
save_plot(name = 'beefjerky')
}
}
\concept{plotutils}
\keyword{internal}
