% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plate.R
\name{printer}
\alias{printer}
\alias{print.96well_plate}
\alias{print.384well_plate}
\alias{print.1536well_plate_t1}
\alias{print.1536well_plate_t2}
\title{Print}
\usage{
\method{print}{`96well_plate`}(plate, ...)

\method{print}{`384well_plate`}(plate, ...)

\method{print}{`1536well_plate_t1`}(plate, ...)

\method{print}{`1536well_plate_t2`}(plate, ...)
}
\arguments{
\item{plate}{plate}

\item{...}{placeholder}
}
\value{
plate

plate

plate

plate
}
\description{
Print
}
\examples{
\dontrun{plate}

}
\concept{printer}
