% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dats.R
\name{read_dats}
\alias{read_dats}
\title{Title: A function to read multiple DAT files into the R environment}
\usage{
read_dats(dat_list)
}
\arguments{
\item{dat_list}{A DAT list (usually list.files() if DAT files are found in the pwd)}
}
\value{
Returns DAT data frames to the R environment
}
\description{
A function that facilitates reading multiple DAT files at once (Some lazy loading)
}
\examples{
fpath <- system.file("extdata", package = "normfluodbf", mustWork = TRUE)
read_dats(fpath)
}
\seealso{
\code{\link[=read_dbfs]{read_dbfs()}}
}
\author{
Tingwei Adeck
}
