% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dbfs.R
\name{read_dbfs}
\alias{read_dbfs}
\title{Title: A function to read multiple dbfs into the R environment}
\usage{
read_dbfs(dbf_list)
}
\arguments{
\item{packages}{A dbf list (usually list.files() if dbf files are found in the pwd)}
}
\value{
Returns DBF data frames to the R environment
}
\description{
A function that facilitates reading multiple dbfs at once (Some lazy loading)
}
\examples{
read_dbfs(list.files())
}
\seealso{
\code{\link[=check_dbfs_pwd]{check_dbfs_pwd()}}
}
\author{
Tingwei Adeck
}
