% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_utils.R
\name{remove_leading_zero}
\alias{remove_leading_zero}
\title{Format Well Names}
\usage{
remove_leading_zero(names_vector)
}
\arguments{
\item{names_vector}{column names}
}
\value{
vector
}
\description{
Format Well Names
}
\examples{
\dontrun{remove_leading_zero(names)}
}
\seealso{
Other plate_utils: 
\code{\link{check_dirt}()},
\code{\link{get_wells_used}()},
\code{\link{plate_data_summary}()},
\code{\link{platename}},
\code{\link{saveloadutils}},
\code{\link{set_plate_version}()},
\code{\link{type}()}
}
\concept{plate_utils}
