% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{removeoutliers}
\alias{removeoutliers}
\alias{remove_outliers}
\alias{remove_outliers.default}
\alias{remove_outliers.96well_plate}
\alias{remove_outliers.384well_plate}
\alias{remove_outliers.1536well_plate_t1}
\alias{remove_outliers.1536well_plate_t2}
\title{Outliers}
\usage{
remove_outliers(plate)

\method{remove_outliers}{default}(plate)

\method{remove_outliers}{`96well_plate`}(plate)

\method{remove_outliers}{`384well_plate`}(plate)

\method{remove_outliers}{`1536well_plate_t1`}(plate)

\method{remove_outliers}{`1536well_plate_t2`}(plate)
}
\arguments{
\item{plate}{plate}
}
\value{
plate

plate

plate

plate

plate

plate

plate
}
\description{
Outliers
}
\note{
Works on a data frame not in well format.

Works on a data frame not in well format.

Works on a data frame not in well format.

Works on a data frame not in well format.

Works on a data frame not in well format.
}
\examples{
\dontrun{remove_outliers(plate)}
}
\concept{removeoutliers}
