% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_word_in_file}
\alias{replace_word_in_file}
\title{Replace Word}
\usage{
replace_word_in_file(file_path, old_word, new_word)
}
\arguments{
\item{file_path}{path}

\item{old_word}{old func name}

\item{new_word}{new func name}
}
\value{
file
}
\description{
Replace Word
}
\note{
Solves the inconvenient problem of renaming a function correctly and having to manually correct it.
}
\examples{
\dontrun{replace_word_in_file('R/plate_plot.R','plot_fluostar_style', 'plot_in_well')}
}
