% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_dat.R
\name{resample_dat}
\alias{resample_dat}
\title{Title: A function to create an attribute or column for each sample loaded into the microplate wells.}
\usage{
resample_dat(df, tnp, cycles)
}
\arguments{
\item{df}{A clean data frame with attributes or tuples containing a mixture of samples.}

\item{tnp}{A numeric value indicating the number of rows used. TNP is used as an acronym for Test, Negative, Positive.}

\item{cycles}{A numeric value indicating the number of cycles selected by the user when running the FLUOstar instrument.}
}
\value{
A new data frame where separated samples are assigned a separate attribute or column.
}
\description{
Designed as a prototype function to take a single attribute or column
consisting of several samples and perform a putative resampling
to yield another data frame with a separate attribute for each sample.
}
\examples{
fpath <- system.file("extdata", "dat_1.dat", package = "normfluodbf", mustWork = TRUE)
dat_df <- read.table(file=fpath)
nocomma_dat <- clean_odddat_optimus(dat_df)
col_1 <- nocomma_dat[,1]
col_1 <- as.data.frame(col_1)
samples_delineated <- resample_dat(col_1, tnp=3, cycles=40)
}
\seealso{
\code{\link[=resample_dat_scale]{resample_dat_scale()}}, \code{\link[=resample_dat_scale_optimus]{resample_dat_scale_optimus()}}
}
\author{
Tingwei Adeck
}
