% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_dat_scale_alt.R
\name{resample_dat_scale_alt}
\alias{resample_dat_scale_alt}
\title{Title: A scaling up of the resample_dat() function}
\usage{
resample_dat_scale_alt(df, tnp, cycles, samples_per_tnp = NULL)
}
\arguments{
\item{df}{A clean data frame with n number of rows}

\item{tnp}{Stands for test,negative,positive (sample types); the number should match the number of sample types in the plate reader even if repeating a sample type}

\item{cycles}{The number of cycles chosen by the researcher. In the case of this package 40 is the standard but ensure to have the right number of samples}
}
\value{
A new data frame with attributes matching the number of sample types and tuples matching the number of cycles. In short it returns delineated samples.
}
\description{
Performs the role of the prototype resample_dat() function but it is scaled up to work on all columns in longer-form data frame.
}
\note{
This function is the pre-requisite to the parent or main function of the update.
As a matter of fact, this function is modified to produce the parent or main function.
}
\examples{
fpath <- system.file("extdata", "dat_1.dat", package = "normfluodbf", mustWork = TRUE)
dat_df <- read.table(file=fpath)
nocomma_dat <- clean_odd_cc(dat_df)
resampled_scaled <- resample_dat_scale_alt(nocomma_dat, tnp=3, cycles=40)
}
\seealso{
\code{\link[=resample_dat]{resample_dat()}}
}
\author{
Tingwei Adeck
}
