% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntdbf_subordinates.R
\name{roundfluor}
\alias{roundfluor}
\title{Title: A value rounding function.}
\usage{
roundfluor(x)
}
\arguments{
\item{x}{Attribute value(s).}
}
\value{
A rounded value with three decimal places when applied to a single value or an attribute with log-transformed values.
}
\description{
Round attribute values to three decimal places.
}
\examples{
test_df <- as.data.frame(c(seq(40)))
colnames(test_df) <- "test"
test_df_norm <- lapply(test_df[1:ncol(test_df)], roundfluor)
}
