% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_demo}
\alias{run_demo}
\title{Run the shiny App
In addition to the functions provided in this package, the \code{normfluodbf} package
also provides an interactive tool that can be used to analyze liposome flux assay data
more easily. The tool will be launched in a web browser.}
\usage{
run_demo()
}
\description{
Run the shiny App
In addition to the functions provided in this package, the \code{normfluodbf} package
also provides an interactive tool that can be used to analyze liposome flux assay data
more easily. The tool will be launched in a web browser.
}
