% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{run_demo_in_background}
\alias{run_demo_in_background}
\title{Run Demo in Background}
\usage{
run_demo_in_background(
  appDir = system.file("shiny/demo", package = "normfluodbf"),
  job_name,
  host,
  port
)
}
\arguments{
\item{appDir}{dir}

\item{job_name}{job name}

\item{host}{host ip (localhost)}

\item{port}{port}
}
\value{
NULL (run script)
}
\description{
Run Demo in Background
}
