% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{run_demo_script}
\alias{run_demo_script}
\title{Run Demo Script}
\usage{
run_demo_script(
  appDir = system.file("shiny/demo", package = "normfluodbf"),
  port,
  host
)
}
\arguments{
\item{appDir}{App Dir}

\item{port}{Port}

\item{host}{Host}
}
\value{
Script
}
\description{
Run Demo Script
}
\details{
A quick script inspired by gptstudio to aid in running the cool normfluodbf demo.
}
\examples{
\dontrun {
run_demo_script (...)}
}
