% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampledata}
\alias{sampledata}
\alias{sample_data_dir}
\alias{sample_data_file}
\title{Get Development Data}
\usage{
sample_data_dir()

sample_data_file(gotofile = NULL)
}
\arguments{
\item{gotofile}{file}
}
\value{
directory data
}
\description{
Get Development Data
}
\examples{
\dontrun{
fpath <- system.file("extdata", package = "normfluodbf", mustWork = TRUE)
sample_data_dir()
sample_data_file(gotofile = NULL)}
}
\seealso{
Other dirutils: 
\code{\link{dirutils}},
\code{\link{normfluodbfcomms}}
}
\concept{dirutils}
