% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_utils.R
\name{saveloadutils}
\alias{saveloadutils}
\alias{save_plate}
\alias{save_rds_plate}
\alias{load_rds_plate}
\alias{var_str}
\title{Save and Load Plate}
\usage{
save_plate(plate, suffix = NULL, interactive = F)

save_rds_plate(plate, save_name, use_tempfile = F)

load_rds_plate(plate, interactive = F)

var_str(var)
}
\arguments{
\item{plate}{plate}

\item{suffix}{suffix}

\item{interactive}{boolean}

\item{save_name}{name}

\item{use_tempfile}{boolean}

\item{var}{variable}
}
\value{
plate

plate

plate

plate
}
\description{
Save and Load Plate
}
\seealso{
Other plate_utils: 
\code{\link{check_dirt}()},
\code{\link{get_wells_used}()},
\code{\link{plate_data_summary}()},
\code{\link{platename}},
\code{\link{remove_leading_zero}()},
\code{\link{set_plate_version}()},
\code{\link{type}()}
}
\concept{plate_utils}
