% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_plate_type.R
\name{set_plate_type}
\alias{set_plate_type}
\title{Set The Child Plate Type}
\usage{
set_plate_type(parent_plate, type)
}
\arguments{
\item{parent_plate}{plate}

\item{type}{child plate type}
}
\value{
plate
}
\description{
Set The Child Plate Type
}
\details{
This is a recursive method.
}
\examples{
\dontrun{
x = set_plate_type(parent_plate,"96well_plate")
print(class(x))
}
}
