% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_utils.R
\name{set_plate_version}
\alias{set_plate_version}
\title{Plate Version}
\usage{
set_plate_version(plate, pkg)
}
\arguments{
\item{plate}{plate}

\item{pkg}{package}
}
\value{
plate
}
\description{
Plate Version
}
\examples{
\dontrun{set_plate_version(plate,pkg)}
}
\seealso{
Other plate_utils: 
\code{\link{check_dirt}()},
\code{\link{get_wells_used}()},
\code{\link{plate_data_summary}()},
\code{\link{platename}},
\code{\link{remove_leading_zero}()},
\code{\link{saveloadutils}},
\code{\link{type}()}
}
\concept{plate_utils}
