% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_params.R
\name{setparams}
\alias{setparams}
\alias{params}
\alias{params<-}
\alias{.params}
\title{Set Plate Params}
\usage{
params(plate, category, name)

params(plate, category, name) <- value

.params(plate, category, name)
}
\arguments{
\item{plate}{plate}

\item{category}{category}

\item{name}{name}
}
\value{
plate

plate

plate
}
\description{
Set Plate Params
}
\examples{
\dontrun{params(plate, cat, name)}

}
\concept{params}
\keyword{internal}
