% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_steps.R
\name{setsteps}
\alias{setsteps}
\alias{steps}
\alias{steps<-}
\title{Set Plate Steps}
\usage{
steps(plate)

steps(plate) <- value
}
\arguments{
\item{plate}{plate}

\item{value}{value}
}
\value{
plate

plate

plate
}
\description{
Set Plate Steps
}
\examples{
\dontrun{steps(plate)}
}
\seealso{
Other steps: 
\code{\link{stepsutils}}
}
\concept{steps}
