% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{source_files}
\alias{source_files}
\title{Title: Source files}
\usage{
source_files(path)
}
\arguments{
\item{path}{path}
}
\value{
All functions from the R file
}
\description{
Title: Source files
}
\keyword{internal}
