% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_status.R
\name{status}
\alias{status}
\alias{status<-}
\alias{dirty}
\alias{dirty<-}
\alias{is_plate_dirty}
\title{Status}
\usage{
status(plate)

status(plate) <- value

dirty(plate)

dirty(plate) <- value

is_plate_dirty(plate)
}
\arguments{
\item{plate}{plate}

\item{value}{value}
}
\value{
plate

plate

plate

plate

plate

logical
}
\description{
Status
}
\examples{
\dontrun{status(plate, status)}

}
\concept{status}
