% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_steps.R
\name{stepsutils}
\alias{stepsutils}
\alias{step}
\alias{step_name}
\alias{.step_name}
\alias{get_step_key_by_index}
\alias{step_begin}
\alias{step_end}
\alias{has_step}
\alias{check_step}
\alias{steps_complete}
\title{Steps Utils}
\usage{
step(plate, step)

step_name(plate, step)

.step_name(plate, step)

get_step_key_by_index(steps, index)

step_begin(...)

step_end(...)

has_step(plate, step)

check_step(plate, step)

steps_complete(plate)
}
\arguments{
\item{plate}{plate}

\item{step}{step}

\item{steps}{steps}

\item{index}{index}

\item{...}{dots}
}
\value{
plate

step number

step name

step name

step name

boolean

boolean

boolean
}
\description{
Steps Utils
}
\note{
Step start time utilizing the cache

Step start time utilizing the cache
}
\examples{
\dontrun{step(plate, step)}
}
\seealso{
Other steps: 
\code{\link{setsteps}}
}
\concept{steps}
