% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset}
\alias{subset}
\alias{subset.default}
\alias{subset.mod_normfluodbf_data}
\alias{subset.normfluodbf_plate}
\alias{subset.96well_plate}
\alias{subset.384well_plate}
\alias{subset.1536well_plate_t1}
\alias{subset.1536well_plate_t2}
\alias{subset_or_not}
\alias{remove_subset_data}
\alias{subset_unplated_data}
\alias{subset_for_layout}
\title{Subset}
\usage{
subset(plate, wells, ...)

\method{subset}{default}(plate, wells, ...)

\method{subset}{mod_normfluodbf_data}(plate, wells, ...)

\method{subset}{normfluodbf_plate}(plate, wells, ...)

\method{subset}{`96well_plate`}(plate, wells, ...)

\method{subset}{`384well_plate`}(plate, wells, ...)

\method{subset}{`1536well_plate_t1`}(plate, wells, ...)

\method{subset}{`1536well_plate_t2`}(plate, wells, ...)

subset_or_not(plate)

remove_subset_data(plate)

subset_unplated_data(data, wells)

subset_for_layout(plate, wells = NULL)
}
\arguments{
\item{plate}{plate}

\item{wells}{wells}

\item{...}{dots}
}
\value{
plate

plate

plate

plate

plate

plate

plate

plate

plate

plot object

plot object

plate

data
}
\description{
Subset
}
\examples{
\dontrun{subset(plate,wells)}
}
\seealso{
Other subsetutils: 
\code{\link{subsetutils}}
}
\concept{subsetutils}
\keyword{internal}
