% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_utils.R
\name{subsetutils}
\alias{subsetutils}
\alias{get_single_well}
\alias{get_wells_btwn}
\alias{.get_wells_btwn}
\alias{range_to_endpoints}
\alias{range_list_to_vec}
\alias{is_range}
\title{Subset Utils}
\usage{
get_single_well(plate, well_id)

get_wells_btwn(well1, well2)

.get_wells_btwn(well1, well2)

range_to_endpoints(well_range)

range_list_to_vec(rangel)

is_range(x)
}
\arguments{
\item{plate}{plate}

\item{well_id}{well id}

\item{well1}{well1}

\item{well2}{well2}

\item{well_range}{well range}

\item{rangel}{rangel}

\item{x}{x}
}
\value{
single well

wells between

wells between

endpoints

wells

boolean
}
\description{
Subset Utils
}
\examples{
\dontrun{get_single_well(plate,well_id)
get_wells_btwn(well1,well2)
range_to_endpoints(well_range = ("A01, B02:C04, C07"))
range_list_to_vec(rangel)}
}
\seealso{
Other subsetutils: 
\code{\link{subset}()}
}
\concept{subsetutils}
\keyword{internal}
