% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_boilerplate}
\alias{test_boilerplate}
\alias{open_testfile}
\title{Test Boilerplate}
\usage{
test_boilerplate(file_name = NULL)

open_testfile(testfile)
}
\arguments{
\item{file_name}{file name}

\item{testfile}{test file}
}
\value{
test boilerplate

open test file
}
\description{
Test Boilerplate

Open Testfile
}
\note{
Solves the inconvenient process of navigating to the tests folder every time
}
\examples{
\dontrun(test_boilerplate(file_name = "test_remove_shit.R"))
\dontrun(open_testfile('test_pipeline.R'))
}
