% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_attribute.R
\name{time_attribute}
\alias{time_attribute}
\title{Title: A time attribute generating function}
\usage{
time_attribute(
  interval = NULL,
  first_end = NULL,
  pause_duration = NULL,
  end_time = NULL,
  cycles = NULL
)
}
\arguments{
\item{interval}{The time interval chosen for the assay often in seconds.}

\item{first_end}{The end time of the initial run, often the pause for the introduction of a new substance. This can be the cycle number chosen for the initial stop.}

\item{pause_duration}{The time between the first end (pause) and resumption of the assay.}

\item{end_time}{The final end time of the assay.}

\item{cycles}{The number of cycles in the assay as selected by the user or researcher.}
}
\value{
The time attribute.
}
\description{
A function to help the researcher reproduce the time attribute.
}
\note{
The original function had an option for minutes which was for less time conscious people
but the final version for this package has no such option. Users MUST provide numbers in
seconds.
}
\examples{
time_test = time_attribute(30,8,136,1276,40)
time_test = time_attribute(60,8,136,2460,40)
}
