% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_utils.R
\name{type}
\alias{type}
\title{Plate Type}
\usage{
type(plate, all = FALSE)
}
\arguments{
\item{plate}{plate}

\item{all}{Boolean}
}
\value{
class attribute
}
\description{
Plate Type
}
\examples{
\dontrun{type(plate)}
}
\seealso{
Other plate_utils: 
\code{\link{check_dirt}()},
\code{\link{get_wells_used}()},
\code{\link{plate_data_summary}()},
\code{\link{platename}},
\code{\link{remove_leading_zero}()},
\code{\link{saveloadutils}},
\code{\link{set_plate_version}()}
}
\concept{plate_utils}
