% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{viewerpane_background_normfluodbf}
\alias{viewerpane_background_normfluodbf}
\title{View App in Viewer Pane}
\usage{
viewerpane_background_normfluodbf(host, port)
}
\arguments{
\item{host}{host}

\item{port}{port}
}
\description{
View App in Viewer Pane
}
