% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{wait_for_bg_shinyapp}
\alias{wait_for_bg_shinyapp}
\title{Validate Url}
\usage{
wait_for_bg_shinyapp(url)
}
\arguments{
\item{url}{url}
}
\description{
Validate Url
}
\examples{
\dontrun {
wait_for_bg_shinyapp(url)}
}
