% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_xy_variables.R
\name{xycoordinates}
\alias{xycoordinates}
\alias{x_var_one}
\alias{x_var_one<-}
\alias{x_var_two}
\alias{x_var_two<-}
\alias{y_var}
\alias{y_var<-}
\alias{x_var_one_label}
\alias{x_var_one_label<-}
\alias{x_var_two_label}
\alias{x_var_two_label<-}
\title{Plot Coordinates}
\usage{
x_var_one(plate)

x_var_one(plate) <- value

x_var_two(plate)

x_var_two(plate) <- value

y_var(plate)

y_var(plate) <- value

x_var_one_label(plate)

x_var_one_label(plate) <- value

x_var_two_label(plate)

x_var_two_label(plate) <- value
}
\arguments{
\item{plate}{plate}

\item{value}{value}
}
\value{
plate

plate

plate

plate

plate

plate

plate

plate

plate

plate
}
\description{
Plot Coordinates
}
\examples{
\dontrun{x_var_one(plate,value)}
}
\concept{xycoordinates}
