% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_evendat.R
\name{clean_evendat}
\alias{clean_evendat}
\title{Title: A function to partially clean dat files with an identified pattern from FLuostar experiments}
\usage{
clean_evendat(df)
}
\arguments{
\item{df}{A data frame with n number of rows}
}
\value{
A new data frame with most impurities taken out. See the example
}
\description{
The function takes a dirty data frame and returns a clean data frame with the only exception being that most values in the returned data frame have commas at the end
The data frame that goes into the function comes from reading dat files from FLUOstar experiments
See the example code
}
\note{
This function only works on a specific type of Fluostar dat file and really serves to be made available for anyone to find a universal utility for the function.
}
\examples{
fpath <- system.file("extdata", "dat_1.dat", package = "normfluodbf", mustWork = TRUE)
dat1_df <- read.table(file=fpath)
partial_cleaned_dat <- clean_evendat(dat1_df)
}
\author{
Tingwei Adeck
}
