% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_odd_cc.R
\name{clean_odd_cc}
\alias{clean_odd_cc}
\title{Title: A function to completely clean dat files from FLuostar experiments}
\usage{
clean_odd_cc(df)
}
\arguments{
\item{df}{A data frame with n number of rows}
}
\value{
A new data frame with NAs retained.
}
\description{
The function takes a dirty data frame and returns a clean data frame.
The function takes care of more QC concerns not seen in previous cleaning functions.
The function accounts for possible machine data formatting when users skip wells by
retaining NA values, which ensures that the user does not loose samples in data analysis.
}
\note{
This function is superior to clean_odddat().
}
\examples{
fpath <- system.file("extdata", "dat_1.dat", package = "normfluodbf", mustWork = TRUE)
dat_df <- read.table(file=fpath)
partial_cleaned_dat <- clean_odddat(dat_df)
}
\seealso{
\code{\link[=clean_odddat]{clean_odddat()}}, \code{\link[=clean_evendat]{clean_evendat()}}, \code{\link[=normfluodat]{normfluodat()}}
}
\author{
Tingwei Adeck
}
