% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma_cleaner.R
\name{comma_cleaner}
\alias{comma_cleaner}
\title{Title: A function to remove unwanted commas in a partially cleaned data frame}
\usage{
comma_cleaner(comma_df)
}
\arguments{
\item{comma_df}{A dirty data frame (numeric values have commas at the end) with n number of rows}
}
\value{
A new data frame with commas at the end of numbers taken out
}
\description{
The function takes a dirty data frame (comma_df) derived from either the clean_odddat or clean_evendat functions.
The resulting data frame is a cleaned data frame with numeric values.
}
\note{
This function is a subordinate function and follows a sequence of actions. In this package, this function cannot be used as a standalone.
}
\examples{
fpath <- system.file("extdata", "dat_1.dat", package = "normfluodbf", mustWork = TRUE)
dat_df <- read.table(file=fpath)
comma_dat <- clean_odddat(dat_df)
nocomma_dat <- comma_cleaner(comma_dat)
}
\author{
Tingwei Adeck
}
