% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntdbf_subordinates.R
\name{decimal_scaling}
\alias{decimal_scaling}
\title{Title: A decimal scaling function (a machine learning tool).}
\usage{
decimal_scaling(x)
}
\arguments{
\item{x}{Attribute value(s).}
}
\value{
A decimal scaled value when applied to a single value or a decimal scaled attribute(s).
}
\description{
Title: A decimal scaling function (a machine learning tool).
}
\note{
The lapply function is required to apply this function across several attributes.
This is NOT a normalization function, so data obtained from the decimal scaling function
exists on a sliding scale and SHOULD NOT be used for meaningful analysis.
}
\examples{
test_df <- as.data.frame(c(seq(40)))
colnames(test_df) <- "test"
test_df_norm <- lapply(test_df[1:ncol(test_df)], decimal_scaling)
}
\references{
https://www.statology.org/how-to-normalize-data-in-r/
}
\author{
Tingwei Adeck
}
