% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluor_threshold_check.R
\name{fluor_threshold_check}
\alias{fluor_threshold_check}
\title{Title: Fluorescence experiment quality control function-Check fluorescence threshold}
\usage{
fluor_threshold_check(clean_df, fun = NA)
}
\arguments{
\item{clean_df}{A cleaned dat or dbf file}

\item{fun}{Used for boolean expressions}
}
\value{
A polite warning message for the researchers next experimental design and the rows and columns with problem values.
}
\description{
Check that fluorescence does not exceed 2^16.
Experimental design issues should be investigated at fluorescence levels this high.
The inventor of this program made a similar mistake when he began these experiments.
}
\examples{
fpath <- system.file("extdata", "dat_1.dat", package = "normfluodbf", mustWork = TRUE)
dat_df <- read.table(file=fpath)
nocomma_dat <- clean_odd_cc(dat_df)
resampled_scaled <- resample_dat_scale(nocomma_dat, tnp=3, cycles=40)
resampled_scaled <- resampled_scaled[,c(1:4)]
fluor_threshold_check(resampled_scaled)
}
