% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_tidy_dbf.R
\name{generic_identifier}
\alias{generic_identifier}
\title{Title: A generic identifier similar to unique identifier but end users supply a column name.}
\usage{
generic_identifier(numrows, col_name)
}
\arguments{
\item{numrows}{The number of rows in a data frame of interest (nrows(df) can be used).}

\item{col_name}{The desired column name for the column.}
}
\value{
A new single column data frame with the desired attribute added.
}
\description{
Title: A generic identifier similar to unique identifier but end users supply a column name.
}
\examples{
generic_identifier(40,col_name="Cycle_No")
}
\author{
Tingwei Adeck
}
