% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntdbf_subordinates.R
\name{min_max_norm}
\alias{min_max_norm}
\title{Title: A function that performs Min-Max normalization(0-1 range) of attributes that require normalization}
\usage{
min_max_norm(x)
}
\arguments{
\item{x}{Column or attribute values passed into the min-max normalization function}
}
\value{
A normalized value (between 0 and 1) when used as a standalone function, or a normalized attribute(s) when used with lapply.
}
\description{
Title: A function that performs Min-Max normalization(0-1 range) of attributes that require normalization
}
\note{
The lapply function is required to apply the function across several columns in a data set.
}
\examples{
test_df <- as.data.frame(c(seq(40)))
colnames(test_df) <- "test"
test_df_norm <- lapply(test_df[1:ncol(test_df)], min_max_norm)
}
\references{
https://www.statology.org/how-to-normalize-data-in-r/
}
\author{
Tingwei Adeck (Adapted from Statology)
}
