% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntdbf_subordinates.R
\name{norm_z}
\alias{norm_z}
\title{Title: A function that performs Z-score standardization (or normalization) of attributes}
\usage{
norm_z(x)
}
\arguments{
\item{x}{Column or attribute values passed into the Z-standardization function}
}
\value{
Attribute(s) with values that have mean 0 and standard deviation 1
}
\description{
Title: A function that performs Z-score standardization (or normalization) of attributes
}
\note{
The lapply function is required to apply the function across several columns in a data set.
}
\examples{
test_df <- as.data.frame(c(seq(40)))
colnames(test_df) <- "test"
test_df_norm <- lapply(test_df[1:ncol(test_df)], norm_z)
}
\references{
https://www.statology.org/how-to-normalize-data-in-r/
}
\author{
Tingwei Adeck
}
