% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normfluodbf-package.R
\docType{package}
\name{normfluodbf-package}
\alias{normfluodbf-package}
\title{normfluodbf: Cleans and Normalizes FLUOstar DBF and DAT Files from Liposome Flux Assays}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Cleans and Normalizes FLUOstar DBF and DAT Files obtained from liposome flux assays. Users should verify extended usage of the package on files from other assay types.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlphaPrime7/normfluodbf}
  \item \url{https://alphaprime7.github.io/normfluodbf/}
  \item Report bugs at \url{https://github.com/AlphaPrime7/normfluodbf/issues}
}

}
\author{
\strong{Maintainer}: Tingwei Adeck \email{awesome.tingwei@outlook.com}

Other contributors:
\itemize{
  \item Tingwei Adeck & Tesla Adeck (\href{https://orcid.org/0009-0009-7450-8863}{ORCID}) [copyright holder]
}

}
\keyword{internal}
