% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{normfluodbfplots}
\alias{normfluodbfplots}
\alias{plot.96well_plate}
\alias{plot.384well_plate}
\alias{plot.1536well_plate_t1}
\alias{plot.1536well_plate_t2}
\title{Plot Plate - Favorite is Fluostar style}
\usage{
\method{plot}{`96well_plate`}(
  x,
  whichplot = 1,
  fluorstarplot = 1 \%in\% whichplot,
  superimpose = 2 \%in\% whichplot,
  plate_layout = 3 \%in\% whichplot,
  plot_side_by_side = 4 \%in\% whichplot,
  legend_labels = NULL,
  plot_name = NULL,
  ...
)

\method{plot}{`384well_plate`}(
  x,
  whichplot = 1,
  fluorstarplot = 1 \%in\% whichplot,
  superimpose = 2 \%in\% whichplot,
  plate_layout = 3 \%in\% whichplot,
  plot_side_by_side = 4 \%in\% whichplot,
  legend_labels = NULL,
  plot_name = NULL,
  ...
)

\method{plot}{`1536well_plate_t1`}(
  x,
  whichplot = 1,
  fluorstarplot = 1 \%in\% whichplot,
  superimpose = 2 \%in\% whichplot,
  plate_layout = 3 \%in\% whichplot,
  plot_side_by_side = 4 \%in\% whichplot,
  legend_labels = NULL,
  plot_name = NULL,
  ...
)

\method{plot}{`1536well_plate_t2`}(
  x,
  whichplot = 1,
  fluorstarplot = 1 \%in\% whichplot,
  superimpose = 2 \%in\% whichplot,
  plate_layout = 3 \%in\% whichplot,
  plot_side_by_side = 4 \%in\% whichplot,
  legend_labels = NULL,
  plot_name = NULL,
  ...
)
}
\arguments{
\item{x}{plot requirement}

\item{whichplot}{int}

\item{fluorstarplot}{whichplot = 1}

\item{superimpose}{whichplot = 2}

\item{plate_layout}{whichplot = 3}

\item{plot_side_by_side}{whichplot = 4}

\item{legend_labels}{labels whichplot = 2,4}

\item{plot_name}{plot name}

\item{...}{additional parameters}
}
\value{
plot object

print plot (return plate)

print plot (return plate)

print plot (return plate)

print plot (return plate)
}
\description{
Plot Plate - Favorite is Fluostar style
}
\examples{
\dontrun{plot(plate, whichplot = 1)}
}
\concept{normfluodbfplots}
