% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_dat.R
\name{resample_dat}
\alias{resample_dat}
\title{Title: A function to extract sample types from the 120 tuples data cleaned data frame.}
\usage{
resample_dat(df, tnp, cycles, samples_per_tnp = NULL)
}
\arguments{
\item{df}{A clean data frame with n number of rows}

\item{tnp}{Stands for test,negative,positive (sample types); the number should match the number of sample types in the plate reader even if repeating a sample type}

\item{cycles}{The number of cycles chosen by the researcher. In the case of this package 40 is the standard but ensure to have the right number of samples}

\item{samples_per_tnp}{An optional parameter thought to be useful but had no use based on the approach taken to solve this problem; Will be useful in future functions}
}
\value{
A new data frame with attributes matching the number of sample types and tuples matching the number of cycles. In short it returns delineated samples.
}
\description{
Designed as a prototype function to take a single column from the cleaned dat data frame
and return the n=3 sample types (n can be any number based on the number of sample types used) as separate attributes with n=40 tuples (n can vary based on the number of cycles ran)
The function is well thought and accounts for almost any scenario.
}
\note{
This function is a subordinate function and prototype that turned out very useful.
}
\examples{
fpath <- system.file("extdata", "dat_1.dat", package = "normfluodbf", mustWork = TRUE)
dat_df <- read.table(file=fpath)
comma_dat <- clean_odddat(dat_df)
nocomma_dat <- comma_cleaner(comma_dat)
col_1 <- nocomma_dat[,1]
col_1 <- as.data.frame(col_1)
samples_delineated <- resample_dat(col_1, tnp=3, cycles=40)
}
\author{
Tingwei Adeck
}
