% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundfluor.R
\name{roundfluor}
\alias{roundfluor}
\title{Title: Round attribute elements or attributes}
\usage{
roundfluor(x)
}
\arguments{
\item{x}{An attribute element that needs to be rounded up}
}
\value{
A rounded up element OR a rounded up attribute when assisted by lapply
}
\description{
Round an attribute element or attribute.
}
\examples{
test_df <- as.data.frame(c(seq(40)))
colnames(test_df) <- "test"
test_df_norm <- lapply(test_df[1:ncol(test_df)], roundfluor)
}
