% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_plate_type.R
\name{set_assiette_type}
\alias{set_assiette_type}
\title{Set The Plate Types}
\usage{
set_assiette_type(plate, type = NULL, child_type = NULL)
}
\arguments{
\item{type}{parent plate type}

\item{child_type}{child plate type}

\item{parent_plate}{A parent plate}
}
\value{
plate
}
\description{
Set The Plate Types
}
\details{
Non-recursive approach.
}
\examples{
\dontrun{set_assiette_type(parent_plate_type,96well_plate)}
}
