% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_attribute.R
\name{time_attribute}
\alias{time_attribute}
\title{Title: A time attribute generating function}
\usage{
time_attribute(
  interval = NULL,
  first_end = NULL,
  pause_duration = NULL,
  end_time = NULL,
  cycles = NULL,
  time_unit = c("cycles", "minutes")
)
}
\arguments{
\item{interval}{The time interval chosen for the assay often in seconds.}

\item{first_end}{The end time of the initial run, often the pause for the introduction of a new substance. This can be the cycle number chosen for the initial stop.}

\item{pause_duration}{The time between the first end (pause) and resumption of the assay.}

\item{end_time}{The final end time of the assay.}

\item{cycles}{The number of cycles in the assay as selected by the user or researcher.}

\item{time_unit}{The time unit used, either cycles or minutes. Works with first_end.}
}
\value{
A rounded value with three decimal places when applied to a single value or an attribute with log-transformed values.
}
\description{
A function to help the researcher reproduce the time attribute.
}
\examples{
time_test = time_attribute(30,8,136,1276,40)
time_test = time_attribute(60,8,136,2460,40)
time_test = time_attribute(0.5,4,2.26,21.26,40,'minutes')
time_test = time_attribute(1,7,2.26,41,40,'minutes')
}
